+++ CHANGELOG.md
@@ -9,8 +9,10 @@ The following is a curated list of changes in the Enact limestone module, newest
 - `limestone/Button` styling to match the latest GUI
 - `limestone/Card` styling to match the latest GUI
 - `limestone/ImageItem` styling to match the latest GUI
+- `limestone/Picker` to use `Steps` instead of own indicator
 - `limestone/ProgressBar` styling to match the latest GUI
 - `limestone/Slider` styling to match the latest GUI
+- `limestone/Steps` styling to match the latest GUI
 - `limestone/TabLayout` styling to match the latest GUI
 
 ### Fixed